Feat/user delete (#5)
* feat: user delete endpoint and logic

* feat: improved removeGuest and added removeHost functions

+++ src/user/user.service.ts
@@ -1,9 +1,10 @@
-import { Injectable } from "@nestjs/common";
+import { Inject, Injectable } from "@nestjs/common";
 import { CreateUserDto } from "./dto/create-user.dto";
 import { UpdateUserDto } from "./dto/update-user.dto";
 import { InjectRepository } from "@nestjs/typeorm";
 import { Repository } from "typeorm";
 import { User } from "./entities/user.entity";
+import { ClientProxy, RpcException } from "@nestjs/microservices";
 import { RpcException } from "@nestjs/microservices";
 import * as bcrypt from "bcrypt";
 import { UpdateCredentialsDto } from "./dto/update-credentials.dto";
@@ -11,6 +12,10 @@ import { UpdateCredentialsDto } from "./dto/update-credentials.dto";
 @Injectable()
 export class UserService {
   constructor(
+    @Inject("RESERVATION_SERVICE") private readonly reservationClient:
+      ClientProxy,
+    @Inject("ACCOMMODATION_SERVICE") private readonly accommodationClient:
+      ClientProxy,
     @InjectRepository(User) private usersRepository: Repository<User>,
   ) {}
   async create(dto: CreateUserDto) {
@@ -24,6 +29,12 @@ export class UserService {
   async findOne(username: string): Promise<User | undefined> {
     return this.usersRepository.findOne(
       { where: { username } },
+    );
+  }
+
+  async findOneId(id: number): Promise<User | undefined> {
+    return this.usersRepository.findOne(
+      { where: { id } },
     );
   }
 
@@ -63,7 +74,22 @@ export class UserService {
     return this.usersRepository.save(user);
   }
 
-  remove(id: number) {
+  removeGuest(id: number) {
+    if (this.reservationClient.send<boolean>("hasActiveReservations", id)) {
+      throw new RpcException(
+        "User cannot be deleted because he has active reservations in future",
+      );
+    }
+    return this.usersRepository.delete(id);
+  }
+
+  removeHost(id: number) {
+    if (this.reservationClient.send<boolean>("hasFutureReservations", id)) {
+      throw new RpcException(
+        "User cannot be deleted because he has future reservations for his accommodations",
+      );
+    }
+    this.accommodationClient.send<any>("deleteHostAccommodations", id);
     return this.usersRepository.delete(id);
   }
